+++ parser/Literal.java
@@ -2,19 +2,19 @@ package parser;
 
 /**
  * A Literal is an AST node that 
- * corresponds to a literal integer value
+ * corresponds to a literal value
  * (a number in the source code).
  */
 public class Literal extends Node {
     
-    private final int value;
+    private final double value;
     
     
     /**
      * Create a new IntLiteral node.
      * @param value the integer value this node evaluates to
      */
-    public Literal(final int value) {
+    public Literal(final double value) {
         super();
         this.value = value;
     }
